Check position of robot, check if robot is at edge
Checks and updates the position of the robot on the board. Checks and turns in the most optimal direction when the robot is at an edge.

This was typed in the GitHub file editor so there may or may not be inconsistencies in variable names or missing brackets. So be wary.